utils.py generated & UserPermissionMixin class created
